Convert esp8266 from Arduino to nodeMCU
With this update we now have the ability to setup the chip without reprogramming! It will create its own AP (ESP Relay #...) with a password of 12345678. The password is necessary otherwise the AP will act weird and not work, this should be left open when I can figure out why open auth works weirdly with nodemcu.

I'm utilizing the blue LED (TX pin) to blink and tell the user that it is in setup mode. This disables printing and showing any html, so an extra led or the current led should be connected to the remaining ESP pin on a new revision. For now, this will do.